requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
flask>=2.0.0
google-generativeai>=0.3.0
Pillow>=9.0.0
